% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_utilities.R
\name{ArcSin}
\alias{ArcSin}
\title{Arc Sine Square Root Transformation for Proportional Data}
\usage{
ArcSin(x)
}
\arguments{
\item{x}{A numerical vector of proportions (must be between 0 and 1).}
}
\description{
This function applies arcsine square root transformation to proportional data.
}
\examples{
ArcSin(runif(100,0,1))

}
\seealso{
\code{\link{SqSin}}
}
\keyword{arcsine,}
\keyword{compositional}
\keyword{data,}
\keyword{proportions,}
\keyword{transformation}
