% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_cv_enet.R
\name{CV.ENET}
\alias{CV.ENET}
\title{Fit a regularized linear model}
\usage{
CV.ENET(metab = metab, metag = metag, alpha = alpha,
  lambda.choice = lambda.choice, nfolds = nfolds, foldid = foldid,
  verbose = verbose, plot = plot, outputDirectory = outputDirectory)
}
\arguments{
\item{metab}{Training data of metabolite relative abundances. 
Should have the exact same rows (subjects/samples) as metag.}

\item{metag}{Training data of microbial sequence features' relative abundances.
Should have the exact same rows (subjects/samples) as metab.}

\item{alpha}{Grid of alpha values between 0 and 1. Default is `seq(0.05, 0.95, 0.05)`.}

\item{lambda.choice}{Choice of optimal lambda ('lambda.min' or 'lambda.1se'). Default is 'lambda.1se'.}

\item{nfolds}{Number of folds for internal cross-validation. Default is 10.}

\item{foldid}{A vector of values between 1 and nfold identifying what fold each observation is in.}

\item{verbose}{Should progress bar be printed. Default is TRUE.}

\item{plot}{Should CV error as a function of lambda be plotted. Default is FALSE.}

\item{outputDirectory}{Name of the desired output directory.}
}
\description{
Fit a regularized linear model
}
\keyword{crobiome,}
\keyword{elastic}
\keyword{metabolite}
\keyword{metabolomics}
\keyword{metagenomics,}
\keyword{mi}
\keyword{net,}
\keyword{prediction,}
