% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_utilities.R
\name{generateLambdaPlot}
\alias{generateLambdaPlot}
\title{Plot CV Error As A Function of Lambda}
\usage{
generateLambdaPlot(CV, x, alpha, outputDirectory)
}
\arguments{
\item{CV}{A glmnet object.}

\item{x}{Index of glmnet fit.}

\item{alpha}{Optimal alpha selected.}

\item{outputDirectory}{Name of the desired output directory. Default is the current directory.}
}
\description{
This function plots the CV error as a function of lambda from Elastic Net Training.
}
\keyword{elastic}
\keyword{metabolite}
\keyword{metabolomics}
\keyword{metagenomics,}
\keyword{microbiome,}
\keyword{net,}
\keyword{prediction,}
