% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_summarize.R
\name{melonnpan.summarize}
\alias{melonnpan.summarize}
\title{Summarize MelonnPan results}
\usage{
melonnpan.summarize(metab = metab, pred = pred, method = method,
  correction = correction, cutoff = cutoff)
}
\arguments{
\item{metab}{Measured metabolite relative abundances (normalized to proportional data ranging from 0 to 1.0). 
Should have same features and samples as pred.}

\item{pred}{Predicted metabolite relative abundances (normalized to proportional data ranging from 0 to 1.0). 
Should have same features and samples as metab.}

\item{method}{Method to correlate measured and predicted metabolites. Default is 'spearman'.}

\item{correction}{Multiplicity adjustment method, same as 'p.adjust'. Default is 'fdr'.}

\item{cutoff}{Q-value threshold for significant prediction. Default is 0.05.}
}
\description{
This function returns prediction accuracy from a pair of measured and predicted metabolites tables.
}
