% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_test_data.R
\docType{data}
\name{melonnpan.test.data}
\alias{melonnpan.test.data}
\title{Normalized gene family abundances in the 
NLIBD dataset (Franzosa et al., 2019).}
\format{A data frame of normalized gene familty abundances, 
as described in Mallick et al. (2019), with subjects in rows and 
UniRef90 IDs in columns.}
\usage{
data(melonnpan.test.data)
}
\description{
Normalized gene family abundance used for 
MelonnPan validation (Mallick et al., 2019).
}
\references{
Franzosa EA et al. (2019). Gut microbiome structure and 
metabolic activity in inflammatory bowel disease. Nature Microbiology 4(2):293-305.

Mallick H, Franzosa EA, McIver LJ, Banerjee S, Sirota-Madi A, 
Kostic AD, Clish CB, Vlamakis H, Xavier R, Huttenhower C (2019). Predictive 
metabolomic profiling of microbial communities using amplicon or 
metagenomic sequences. Nature Communications 10(1):3136-3146.
}
\keyword{datasets}
