% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_summarize.R, R/melonnpan_visualize.R
\name{melonnpan.visualize}
\alias{melonnpan.visualize}
\title{Visualization of Significant Results.}
\usage{
melonnpan.visualize(metab, pred, cohenCorrelationCutoff, Output_file,
  ncol = 2, nrow = 2)

melonnpan.visualize(metab, pred, cohenCorrelationCutoff, Output_file,
  ncol = 2, nrow = 2)
}
\arguments{
\item{metab}{Measured metabolite relative abundances (normalized to proportional data ranging from 0 to 1.0). 
Should have same features and samples as pred.}

\item{pred}{Predicted metabolite relative abundances (normalized to proportional data ranging from 0 to 1.0). 
Should have same features and samples as metab.}

\item{cohenCorrelationCutoff}{Correlation threshold for significant prediction. Default is 0.3.}

\item{Output_file}{Path to the file to write the output.}

\item{ncol}{Number of columns in batch scatter plot. Default is 2.}

\item{nrow}{Number of rows in batch scatter plot. Default is 2.}

\item{metab}{Measured metabolite relative abundances (normalized to proportional data ranging from 0 to 1.0). 
Should have same features and samples as pred.}

\item{pred}{Predicted metabolite relative abundances (normalized to proportional data ranging from 0 to 1.0). 
Should have same features and samples as metab.}

\item{Output_file}{Path to the file to write the output.}

\item{cohenCorrelationCutoff}{Correlation threshold for significant prediction. Default is 0.3.}

\item{nrow}{Number of rows in batch scatter plot. Default is 2.}

\item{ncol}{Number of columns in batch scatter plot. Default is 2.}
}
\description{
This function produces scatter plots for significantly predicted compounds.

This function produces scatter plots for significantly predicted compounds.
}
\keyword{elastic}
\keyword{metabolite}
\keyword{metabolomics}
\keyword{metagenomics,}
\keyword{microbiome,}
\keyword{net,}
\keyword{prediction,}
