% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_utilities.R
\name{readTable}
\alias{readTable}
\title{Read data using a wrapper of data.table}
\usage{
readTable(Input)
}
\arguments{
\item{Input}{Full path of the input data.}
}
\description{
This function reads the input data using the data.table package 
but returns a matrix with row names (unlike fread).
}
\keyword{elastic}
\keyword{metabolite}
\keyword{metabolomics}
\keyword{metagenomics,}
\keyword{microbiome,}
\keyword{net,}
\keyword{prediction,}
