% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_utilities.R
\name{rntransform}
\alias{rntransform}
\title{Rank-based inverse normal tranformation}
\usage{
rntransform(y, split_ties = FALSE)
}
\arguments{
\item{y}{a numeric vector which will be rank normal transformed}

\item{split_ties}{a binary string of FALSE (default) or TRUE indicating if tied values, of the same rank, should be randomly split giving them unique ranks.}
}
\value{
returns a numeric vector, with the same length as y, of rank normal transformed values
}
\description{
This function rank normal transforms a vector of data. The procedure is built off of that provided in the GenABEL pacakge.
}
\examples{
## simulate a negative binomial distribution of values
nb_data = rnbinom(500, mu = 40, size = 100)
## rank normal transform those values
rnt_data = rntransform( nb_data , split_ties = TRUE )

}
\keyword{normal}
\keyword{rank}
\keyword{transformation}
