% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_utilities.R
\name{SqSin}
\alias{SqSin}
\title{Back Transformation of Predicted Values from A Statistical Model with
ArcSine Square Root Transformed Compositional Response Variable}
\usage{
SqSin(x, adjust = TRUE)
}
\arguments{
\item{x}{A numerical vector of arcsine square root transformed compositions or proportions.}

\item{adjust}{Should values less than 0 be mapped to 0 and 
values greater than pi/2 be mapped to 1? Defaults to TRUE.}
}
\description{
This function applies back transformation to the predicted values of a statistical model with
arcsine square root transformed compositional response variable.
}
\examples{

# This is a simple demonstration
y <- ArcSin(runif(100,0,1))
x <- rnorm(100,0,1)
z <- predict(lm(y ~ x))
SqSin(z)
}
\seealso{
\code{\link{ArcSin}}
}
\keyword{arcsine,}
\keyword{back}
\keyword{compositional}
\keyword{data,}
\keyword{proportions,}
\keyword{transformation}
