% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melonnpan_predict.R
\name{melonnpan.predict}
\alias{melonnpan.predict}
\title{Model-based Genomically Informed High-dimensional Predictor 
of Microbial Community Metabolite Profiles}
\usage{
melonnpan.predict(metag, weight.matrix = NULL, train.metag = NULL,
  criticalpoint = 0.9793, corr.method = "pearson", output)
}
\arguments{
\item{metag}{Microbial sequence features' relative abundances (matrix) 
for which prediction is desired. The sequence features' abundances are expected to 
be normalized (i.e. proportional data ranging from 0 to 1.0).}

\item{weight.matrix}{The weight matrix to be used for prediction (optional). 
If not provided, by default, a pre-trained weight matrix based on UniRef90 gene 
families from the original MelonnPan paper (Mallick et al, 2019) will be used.}

\item{train.metag}{Quality-controlled training metagenomes against which 
similarity is desired (optional). The sequence features' abundances are expected 
to be normalized (i.e. proportional data ranging from 0.0 to 1.0).
If not provided, a pre-processed UniRef90 gene family training table from 
the original MelonnPan paper (Mallick et al. 2019) will be used.}

\item{criticalpoint}{A numeric value corresponding to the significance level 
to find the top PCs. If the significance level is 0.05, 0.01, 0.005, or 0.001, 
the criticalpoint should be set to be 0.9793, 2.0234, 2.4224, or 3.2724, 
accordingly. The default is 0.9793 (i.e. 0.05 significance level).}

\item{corr.method}{Method to correlate new metagenomes and training PCs. 
Default is 'pearson'.}

\item{output}{Path to the file to write the output.}
}
\description{
Predict metabolites from new microbiome samples.
}
\keyword{elastic}
\keyword{metabolite}
\keyword{metabolomics}
\keyword{metagenomics,}
\keyword{microbiome,}
\keyword{net,}
\keyword{prediction,}
